# internal/dips/api/gen/http/openapi.json
{
  "swagger": "2.0",
  "info": {
    "title": "DIPs API",
    "description": "The DIPs API is used to request DIP creation and retrieve DIP details.",
    "version": "0.2.0"
  },
  "host": "localhost:8080",
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json",
    "application/xml",
    "application/gob"
  ],
  "paths": {
    "/dips": {
      "post": {
        "tags": [
          "DIPs"
        ],
        "summary": "create DIPs",
        "description": "The create method requests DIP creation for the given document key.",
        "operationId": "DIPs#create",
        "parameters": [
          {
            "name": "object",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "docKey": {
                  "type": "string",
                  "description": "The docKey field contains the document key used to create the DIP.",
                  "example": "CH-000001",
                  "minLength": 1
                },
                "ignoreCache": {
                  "type": "boolean",
                  "description": "The ignoreCache field indicates whether to ignore a cached DIP previously created for this docKey. When ignoreCache is true, a new DIP is created for the given docKey even if a cached DIP exists. When ignoreCache is false or omitted, the ID of a cached DIP may be returned.",
                  "default": false,
                  "example": true
                }
              }
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted response.",
            "schema": {
              "$ref": "#/definitions/DIPsCreateResponseBody",
              "required": [
                "id"
              ]
            }
          },
          "400": {
            "description": "Bad Request response.",
            "schema": {
              "$ref": "#/definitions/DIPsCreateBadRequestResponseBody"
            }
          },
          "401": {
            "description": "Unauthorized response.",
            "schema": {
              "$ref": "#/definitions/DIPsCreateUnauthorizedResponseBody"
            }
          },
          "500": {
            "description": "Internal Server Error response.",
            "schema": {
              "$ref": "#/definitions/DIPsCreateInternalErrorResponseBody"
            }
          },
          "501": {
            "description": "Not Implemented response.",
            "schema": {
              "$ref": "#/definitions/DIPsCreateNotImplementedResponseBody"
            }
          }
        },
        "schemes": [
          "http"
        ],
        "security": [
          {
            "bearer_header_Authorization": null
          }
        ]
      }
    },
    "/dips/{id}": {
      "get": {
        "tags": [
          "DIPs"
        ],
        "summary": "show DIPs",
        "description": "The show method retrieves DIP details.",
        "operationId": "DIPs#show",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "DIPID uniquely identifies a DIP.",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "OK response.",
            "schema": {
              "$ref": "#/definitions/DIPsShowResponseBody",
              "required": [
                "id",
                "docKey",
                "status",
                "created_at"
              ]
            }
          },
          "400": {
            "description": "Bad Request response.",
            "schema": {
              "$ref": "#/definitions/DIPsShowBadRequestResponseBody"
            }
          },
          "401": {
            "description": "Unauthorized response.",
            "schema": {
              "$ref": "#/definitions/DIPsShowUnauthorizedResponseBody"
            }
          },
          "404": {
            "description": "Not Found response.",
            "schema": {
              "$ref": "#/definitions/DIPsShowNotFoundResponseBody"
            }
          },
          "500": {
            "description": "Internal Server Error response.",
            "schema": {
              "$ref": "#/definitions/DIPsShowInternalErrorResponseBody"
            }
          },
          "501": {
            "description": "Not Implemented response.",
            "schema": {
              "$ref": "#/definitions/DIPsShowNotImplementedResponseBody"
            }
          }
        },
        "schemes": [
          "http"
        ],
        "security": [
          {
            "bearer_header_Authorization": null
          }
        ]
      }
    },
...
  },
  "definitions": {
    "DIPsCreateBadRequestResponseBody": {
      "title": "Mediatype identifier: application/vnd.goa.error; view=default",
      "type": "object",
      "properties": {
        "fault": {
          "type": "boolean",
          "description": "Is the error a server-side fault?",
          "example": false
        },
        "id": {
          "type": "string",
          "description": "ID is a unique identifier for this particular occurrence of the problem.",
          "example": "123abc"
        },
        "message": {
          "type": "string",
          "description": "Message is a human-readable explanation specific to this occurrence of the problem.",
          "example": "parameter 'p' must be an integer"
        },
        "name": {
          "type": "string",
          "description": "Name is the name of this class of errors.",
          "example": "bad_request"
        },
        "temporary": {
          "type": "boolean",
          "description": "Is the error temporary?",
          "example": false
        },
        "timeout": {
          "type": "boolean",
          "description": "Is the error a timeout?",
          "example": false
        }
      },
      "description": "Error response result type (default view)",
      "example": {
        "fault": false,
        "id": "123abc",
        "message": "parameter 'p' must be an integer",
        "name": "bad_request",
        "temporary": false,
        "timeout": false
      },
      "required": [
        "name",
        "id",
        "message",
        "temporary",
        "timeout",
        "fault"
      ]
    },
    "DIPsCreateInternalErrorResponseBody": {
      "title": "Mediatype identifier: application/vnd.goa.error; view=default",
      "type": "object",
      "properties": {
        "fault": {
          "type": "boolean",
          "description": "Is the error a server-side fault?",
          "example": false
        },
        "id": {
          "type": "string",
          "description": "ID is a unique identifier for this particular occurrence of the problem.",
          "example": "123abc"
        },
        "message": {
          "type": "string",
          "description": "Message is a human-readable explanation specific to this occurrence of the problem.",
          "example": "parameter 'p' must be an integer"
        },
        "name": {
          "type": "string",
          "description": "Name is the name of this class of errors.",
          "example": "bad_request"
        },
        "temporary": {
          "type": "boolean",
          "description": "Is the error temporary?",
          "example": false
        },
        "timeout": {
          "type": "boolean",
          "description": "Is the error a timeout?",
          "example": false
        }
      },
      "description": "Error response result type (default view)",
      "example": {
        "fault": false,
        "id": "123abc",
        "message": "parameter 'p' must be an integer",
        "name": "bad_request",
        "temporary": false,
        "timeout": false
      },
      "required": [
        "name",
        "id",
        "message",
        "temporary",
        "timeout",
        "fault"
      ]
    },
    "DIPsCreateNotImplementedResponseBody": {
      "title": "Mediatype identifier: application/vnd.goa.error; view=default",
      "type": "object",
      "properties": {
        "fault": {
          "type": "boolean",
          "description": "Is the error a server-side fault?",
          "example": false
        },
        "id": {
          "type": "string",
          "description": "ID is a unique identifier for this particular occurrence of the problem.",
          "example": "123abc"
        },
        "message": {
          "type": "string",
          "description": "Message is a human-readable explanation specific to this occurrence of the problem.",
          "example": "parameter 'p' must be an integer"
        },
        "name": {
          "type": "string",
          "description": "Name is the name of this class of errors.",
          "example": "bad_request"
        },
        "temporary": {
          "type": "boolean",
          "description": "Is the error temporary?",
          "example": false
        },
        "timeout": {
          "type": "boolean",
          "description": "Is the error a timeout?",
          "example": false
        }
      },
      "description": "Error response result type (default view)",
      "example": {
        "fault": false,
        "id": "123abc",
        "message": "parameter 'p' must be an integer",
        "name": "bad_request",
        "temporary": false,
        "timeout": false
      },
      "required": [
        "name",
        "id",
        "message",
        "temporary",
        "timeout",
        "fault"
      ]
    },
    "DIPsCreateResponseBody": {
      "title": "DIPsCreateResponseBody",
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "description": "The id field contains the identifier assigned to the DIP.",
          "example": "3f38d6f4-7b19-4db8-8d7d-693b84a9a2fb",
          "format": "uuid"
        }
      },
      "example": {
        "id": "3f38d6f4-7b19-4db8-8d7d-693b84a9a2fb"
      },
      "required": [
        "id"
      ]
    },
    "DIPsCreateUnauthorizedResponseBody": {
      "title": "Mediatype identifier: application/vnd.goa.error; view=default",
      "type": "object",
      "properties": {
        "fault": {
          "type": "boolean",
          "description": "Is the error a server-side fault?",
          "example": false
        },
        "id": {
          "type": "string",
          "description": "ID is a unique identifier for this particular occurrence of the problem.",
          "example": "123abc"
        },
        "message": {
          "type": "string",
          "description": "Message is a human-readable explanation specific to this occurrence of the problem.",
          "example": "parameter 'p' must be an integer"
        },
        "name": {
          "type": "string",
          "description": "Name is the name of this class of errors.",
          "example": "bad_request"
        },
        "temporary": {
          "type": "boolean",
          "description": "Is the error temporary?",
          "example": false
        },
        "timeout": {
          "type": "boolean",
          "description": "Is the error a timeout?",
          "example": false
        }
      },
      "description": "Error response result type (default view)",
      "example": {
        "fault": false,
        "id": "123abc",
        "message": "parameter 'p' must be an integer",
        "name": "bad_request",
        "temporary": false,
        "timeout": false
      },
      "required": [
        "name",
        "id",
        "message",
        "temporary",
        "timeout",
        "fault"
      ]
    },
...
    "DIPsShowBadRequestResponseBody": {
      "title": "Mediatype identifier: application/vnd.goa.error; view=default",
      "type": "object",
      "properties": {
        "fault": {
          "type": "boolean",
          "description": "Is the error a server-side fault?",
          "example": false
        },
        "id": {
          "type": "string",
          "description": "ID is a unique identifier for this particular occurrence of the problem.",
          "example": "123abc"
        },
        "message": {
          "type": "string",
          "description": "Message is a human-readable explanation specific to this occurrence of the problem.",
          "example": "parameter 'p' must be an integer"
        },
        "name": {
          "type": "string",
          "description": "Name is the name of this class of errors.",
          "example": "bad_request"
        },
        "temporary": {
          "type": "boolean",
          "description": "Is the error temporary?",
          "example": false
        },
        "timeout": {
          "type": "boolean",
          "description": "Is the error a timeout?",
          "example": false
        }
      },
      "description": "Error response result type (default view)",
      "example": {
        "fault": false,
        "id": "123abc",
        "message": "parameter 'p' must be an integer",
        "name": "bad_request",
        "temporary": false,
        "timeout": false
      },
      "required": [
        "name",
        "id",
        "message",
        "temporary",
        "timeout",
        "fault"
      ]
    },
    "DIPsShowInternalErrorResponseBody": {
      "title": "Mediatype identifier: application/vnd.goa.error; view=default",
      "type": "object",
      "properties": {
        "fault": {
          "type": "boolean",
          "description": "Is the error a server-side fault?",
          "example": false
        },
        "id": {
          "type": "string",
          "description": "ID is a unique identifier for this particular occurrence of the problem.",
          "example": "123abc"
        },
        "message": {
          "type": "string",
          "description": "Message is a human-readable explanation specific to this occurrence of the problem.",
          "example": "parameter 'p' must be an integer"
        },
        "name": {
          "type": "string",
          "description": "Name is the name of this class of errors.",
          "example": "bad_request"
        },
        "temporary": {
          "type": "boolean",
          "description": "Is the error temporary?",
          "example": false
        },
        "timeout": {
          "type": "boolean",
          "description": "Is the error a timeout?",
          "example": false
        }
      },
      "description": "Error response result type (default view)",
      "example": {
        "fault": false,
        "id": "123abc",
        "message": "parameter 'p' must be an integer",
        "name": "bad_request",
        "temporary": false,
        "timeout": false
      },
      "required": [
        "name",
        "id",
        "message",
        "temporary",
        "timeout",
        "fault"
      ]
    },
    "DIPsShowNotFoundResponseBody": {
      "title": "Mediatype identifier: application/vnd.goa.error; view=default",
      "type": "object",
      "properties": {
        "fault": {
          "type": "boolean",
          "description": "Is the error a server-side fault?",
          "example": false
        },
        "id": {
          "type": "string",
          "description": "ID is a unique identifier for this particular occurrence of the problem.",
          "example": "123abc"
        },
        "message": {
          "type": "string",
          "description": "Message is a human-readable explanation specific to this occurrence of the problem.",
          "example": "parameter 'p' must be an integer"
        },
        "name": {
          "type": "string",
          "description": "Name is the name of this class of errors.",
          "example": "bad_request"
        },
        "temporary": {
          "type": "boolean",
          "description": "Is the error temporary?",
          "example": false
        },
        "timeout": {
          "type": "boolean",
          "description": "Is the error a timeout?",
          "example": false
        }
      },
      "description": "Error response result type (default view)",
      "example": {
        "fault": false,
        "id": "123abc",
        "message": "parameter 'p' must be an integer",
        "name": "bad_request",
        "temporary": false,
        "timeout": false
      },
      "required": [
        "name",
        "id",
        "message",
        "temporary",
        "timeout",
        "fault"
      ]
    },
    "DIPsShowNotImplementedResponseBody": {
      "title": "Mediatype identifier: application/vnd.goa.error; view=default",
      "type": "object",
      "properties": {
        "fault": {
          "type": "boolean",
          "description": "Is the error a server-side fault?",
          "example": false
        },
        "id": {
          "type": "string",
          "description": "ID is a unique identifier for this particular occurrence of the problem.",
          "example": "123abc"
        },
        "message": {
          "type": "string",
          "description": "Message is a human-readable explanation specific to this occurrence of the problem.",
          "example": "parameter 'p' must be an integer"
        },
        "name": {
          "type": "string",
          "description": "Name is the name of this class of errors.",
          "example": "bad_request"
        },
        "temporary": {
          "type": "boolean",
          "description": "Is the error temporary?",
          "example": false
        },
        "timeout": {
          "type": "boolean",
          "description": "Is the error a timeout?",
          "example": false
        }
      },
      "description": "Error response result type (default view)",
      "example": {
        "fault": false,
        "id": "123abc",
        "message": "parameter 'p' must be an integer",
        "name": "bad_request",
        "temporary": false,
        "timeout": false
      },
      "required": [
        "name",
        "id",
        "message",
        "temporary",
        "timeout",
        "fault"
      ]
    },
    "DIPsShowResponseBody": {
      "title": "DIPsShowResponseBody",
      "type": "object",
      "properties": {
        "completed_at": {
          "type": "string",
          "description": "The completed_at field contains the time when DIP processing completed.",
          "example": "2026-05-27T15:04:05Z",
          "format": "date-time"
        },
        "created_at": {
          "type": "string",
          "description": "The created_at field contains the time when the DIP was requested.",
          "example": "2026-05-27T15:04:05Z",
          "format": "date-time"
        },
        "docKey": {
          "type": "string",
          "description": "The docKey field contains the document key used to create the DIP.",
          "example": "CH-000001",
          "minLength": 1
        },
        "error_message": {
          "type": "string",
          "description": "The error_message field contains an error message if the DIP failed.",
          "example": "ACTAPro returned: 404 Not Found; Document not found"
        },
        "id": {
          "type": "string",
          "description": "The id field uniquely identifies the DIP.",
          "example": "3f38d6f4-7b19-4db8-8d7d-693b84a9a2fb",
          "format": "uuid"
        },
        "object_key": {
          "type": "string",
          "description": "The object_key field contains the object store key for the completed DIP.",
          "example": "dips/3f38d6f4-7b19-4db8-8d7d-693b84a9a2fb.zip",
          "minLength": 1
        },
        "started_at": {
          "type": "string",
          "description": "The started_at field contains the time when DIP processing started.",
          "example": "2026-05-27T15:04:05Z",
          "format": "date-time"
        },
        "status": {
          "type": "string",
          "description": "The status field contains the current DIP status.",
          "example": "done",
          "enum": [
            "queued",
            "in progress",
            "done",
            "failed"
          ]
        }
      },
      "example": {
        "completed_at": "2026-05-27T15:04:05Z",
        "created_at": "2026-05-27T15:04:05Z",
        "docKey": "CH-000001",
        "error_message": "ACTAPro returned: 404 Not Found; Document not found",
        "id": "3f38d6f4-7b19-4db8-8d7d-693b84a9a2fb",
        "object_key": "dips/3f38d6f4-7b19-4db8-8d7d-693b84a9a2fb.zip",
        "started_at": "2026-05-27T15:04:05Z",
        "status": "done"
      },
      "required": [
        "id",
        "docKey",
        "status",
        "created_at"
      ]
    },
    "DIPsShowUnauthorizedResponseBody": {
      "title": "Mediatype identifier: application/vnd.goa.error; view=default",
      "type": "object",
      "properties": {
        "fault": {
          "type": "boolean",
          "description": "Is the error a server-side fault?",
          "example": false
        },
        "id": {
          "type": "string",
          "description": "ID is a unique identifier for this particular occurrence of the problem.",
          "example": "123abc"
        },
        "message": {
          "type": "string",
          "description": "Message is a human-readable explanation specific to this occurrence of the problem.",
          "example": "parameter 'p' must be an integer"
        },
        "name": {
          "type": "string",
          "description": "Name is the name of this class of errors.",
          "example": "bad_request"
        },
        "temporary": {
          "type": "boolean",
          "description": "Is the error temporary?",
          "example": false
        },
        "timeout": {
          "type": "boolean",
          "description": "Is the error a timeout?",
          "example": false
        }
      },
      "description": "Error response result type (default view)",
      "example": {
        "fault": false,
        "id": "123abc",
        "message": "parameter 'p' must be an integer",
        "name": "bad_request",
        "temporary": false,
        "timeout": false
      },
      "required": [
        "name",
        "id",
        "message",
        "temporary",
        "timeout",
        "fault"
      ]
    }
  },
  "securityDefinitions": {
    "bearer_header_Authorization": {
      "type": "apiKey",
      "description": "Authenticate with an access token supplied as Authorization: Bearer [token]. Clients obtain tokens from an external OIDC provider.",
      "name": "Authorization",
      "in": "header"
    }
  }
}